Rewrite add/edit command to use abstract component
