+++ tests/components/integrations/add_command.test.jsx
@@ -5,21 +5,15 @@ import React from 'react';
 
 import {shallow} from 'enzyme';
 
-import * as Utils from 'utils/utils.jsx';
-
 import AddCommand from 'components/integrations/components/add_command/add_command.jsx';
 
 describe('components/integrations/AddCommand', () => {
     test('should match snapshot', () => {
-        function emptyFunction() {} //eslint-disable-line no-empty-function
-        const teamId = Utils.generateId();
+        const emptyFunction = jest.fn();
 
         const wrapper = shallow(
             <AddCommand
-                team={{
-                    id: teamId,
-                    name: 'test'
-                }}
+                team={{name: 'test'}}
                 addCommandRequest={{
                     status: 'not_started',
                     error: null
